+++ tensorflow_datasets/audio/librispeech.py
@@ -22,7 +22,7 @@ from __future__ import print_function
 import collections
 import os
 
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 
 import tensorflow_datasets.public_api as tfds