+++ tensorflow_datasets/image/cars196.py
@@ -20,7 +20,7 @@ from __future__ import division
 from __future__ import print_function
 import os
 import six.moves.urllib as urllib
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 import tensorflow_datasets.public_api as tfds
 
 _URL = 'http://imagenet.stanford.edu/internal/car196/'