+++ tensorflow_datasets/image/cbis_ddsm.py
@@ -25,7 +25,7 @@ import re
 
 from absl import logging
 import numpy as np
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 import tensorflow_datasets.public_api as tfds