+++ tensorflow_datasets/summarization/aeslc.py
@@ -21,7 +21,7 @@ from __future__ import print_function
 
 import os
 
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 import tensorflow_datasets.public_api as tfds
 
 _CITATION = """
@@ -69,7 +69,9 @@ class Aeslc(tfds.core.GeneratorBasedBuilder):
 
   def _split_generators(self, dl_manager):
     """Returns SplitGenerators."""
-    dl_path = dl_manager.download_and_extract(_URL)
+    dl_path = dl_manager.download_and_extract(tfds.download.Resource(
+        url=_URL,
+        extract_method=tfds.download.ExtractMethod.ZIP))
     input_path = os.path.join(dl_path, "AESLC-master", "enron_subject_line")
     return [
         tfds.core.SplitGenerator(