+++ tensorflow_datasets/testing/test_utils_test.py
@@ -19,7 +19,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 
 from tensorflow_datasets.testing import test_case
 from tensorflow_datasets.testing import test_utils