+++ tensorflow_datasets/text/glue.py
@@ -24,7 +24,7 @@ import os
 import numpy as np
 import six
 
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 import tensorflow_datasets.public_api as tfds
 
 _GLUE_CITATION = """\