+++ tensorflow_datasets/text/super_glue.py
@@ -23,7 +23,7 @@ import json
 import os
 
 import six
-import tensorflow as tf
+import tensorflow.compat.v2 as tf
 import tensorflow_datasets.public_api as tfds
 
 _SUPER_GLUE_CITATION = """\