+++ docs/release_notes.md
@@ -17,3 +17,5 @@
     [VTAB benchmark](https://arxiv.org/abs/1910.04867).
 *   Add e-SNLI dataset from the paper
     [e-SNLI](http://papers.nips.cc/paper/8163-e-snli-natural-language-inference-with-natural-language-explanations.pdf).
+*   Add SCAN dataset introduced
+    [here](https://arxiv.org/pdf/1711.00350.pdf).